pyproject.toml
[project]
name = "medi-agent"
version = "0.1.0"
description = "medi_agent is a medical chatbot"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "rich (>=13.9.4,<14.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)",
    "alembic (>=1.15.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
